# docs/swagger.yaml
basePath: /
definitions:
  web.JabatanTimCreateRequest:
    properties:
      level_jabatan:
        type: integer
      nama_jabatan:
        type: string
    type: object
  web.JabatanTimResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      level_jabatan:
        type: integer
      nama_jabatan:
        type: string
      updated_at:
        type: string
    type: object
  web.JabatanTimUpdateRequest:
    properties:
      id:
        type: integer
      level_jabatan:
        type: integer
      nama_jabatan:
        type: string
    type: object
  web.SusunanTimCreateRequest:
    properties:
      is_active:
        type: boolean
      keterangan:
        type: string
      kode_tim:
        type: string
      nama_jabatan_tim:
        type: string
      nip:
        type: string
    type: object
...
  web.SusunanTimResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      keterangan:
        type: string
      kode_tim:
        type: string
      nama_jabatan_tim:
        type: string
      nip:
        type: string
      updated_at:
        type: string
    type: object
  web.SusunanTimUpdateRequest:
    properties:
      id:
        type: integer
      is_active:
        type: boolean
      keterangan:
        type: string
      kode_tim:
        type: string
      nama_jabatan_tim:
        type: string
      nip:
        type: string
    type: object
...
  web.WebResponse:
    description: Web Response
    properties:
      code:
        type: integer
      data: {}
      status:
        type: string
    type: object
host: ${HOST}:${PORT}
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API For Tim Kerja Services
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Tim Kerja Service API
  version: "1.0"
paths:
  /jabatantim:
    get:
      consumes:
      - application/json
      description: Get All Jabatan Tim
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/web.JabatanTimResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Get All Jabatan Tim
      tags:
      - Jabatan Tim
    post:
      consumes:
      - application/json
      description: Create new Jabatan Tim
      parameters:
      - description: Jabatan Tim Create Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.JabatanTimCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.JabatanTimResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Create Jabatan Tim
      tags:
      - Jabatan Tim
  /jabatantim/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Jabatan Tim
      parameters:
      - description: Jabatan Tim ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Delete Jabatan Tim
      tags:
      - Jabatan Tim
    get:
      consumes:
      - application/json
      description: Get Jabatan Tim detail by ID
      parameters:
      - description: Jabatan Tim ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.JabatanTimResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Get Jabatan Tim by ID
      tags:
      - Jabatan Tim
    put:
      consumes:
      - application/json
      description: Update Jabatan Tim
      parameters:
      - description: Jabatan Tim ID
        in: path
        name: id
        required: true
        type: integer
      - description: Jabatan Tim Update Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.JabatanTimUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.JabatanTimResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Update Jabatan Tim
      tags:
      - Jabatan Tim
...
  /susunantim:
    get:
      consumes:
      - application/json
      description: Get All Susunan Tim
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/web.SusunanTimResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Get All Susunan Tim
      tags:
      - Susunan Tim
    post:
      consumes:
      - application/json
      description: Create new Susunan Tim
      parameters:
      - description: Susunan Tim Create Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.SusunanTimCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.SusunanTimResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Create Susunan Tim
      tags:
      - Susunan Tim
  /susunantim/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Susunan Tim
      parameters:
      - description: Susunan Tim ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Delete Susunan Tim
      tags:
      - Susunan Tim
    get:
      consumes:
      - application/json
      description: Get Susunan Tim detail by ID
      parameters:
      - description: Susunan Tim ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.SusunanTimResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Get Susunan Tim by ID
      tags:
      - Susunan Tim
    put:
      consumes:
      - application/json
      description: Update Susunan Tim
      parameters:
      - description: Susunan Tim ID
        in: path
        name: id
        required: true
        type: integer
      - description: Susunan Tim Update Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.SusunanTimUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.SusunanTimResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Update Susunan Tim
      tags:
      - Susunan Tim
...
swagger: "2.0"